refactor: folder structure for improved organization
- Moved connectivity status and home active screen enums to `src/domain/model/`.
- Updated imports across affected files to reflect new paths.
- Enhanced clarity and maintainability of the project structure.
