+++ lib/src/domain/model/connectivity_status.dart
