+++ test/src/services/mixins/mosque_helpers_mixins_test.dart
@@ -2,7 +2,7 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mawaqit/src/enum/home_active_screen.dart';
+import 'package:mawaqit/src/domain/model/home_active_screen.dart';
 import 'package:mawaqit/src/models/mosque.dart';
 import 'package:mawaqit/src/models/mosqueConfig.dart';
 import 'package:mawaqit/src/models/times.dart';
